+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move replication-related modules from `core` module under `core::replication`.
 - Move `Replicated` to the `replication` module.
 - Split the `ctx` module and move event-related contexts under `core::events_registry::ctx` and replication-related contexts under `core::replication_registry::ctx`.
+- Separate paths from `diagnostics` module by `/` and their parent path now `client/replication` instead of `replication/client`.
 - Rename `ServerPlugin::change_timeout` into `ServerPlugin::mutations_timeout`.
 - Rename `ServerInitTick` into `ServerChangeTick`.
 - Rename `ReplicatedClient::init_tick` into `ReplicatedClient::change_tick`.